+++ CHANGELOG.md
@@ -8,6 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### v0.4.0
 ### Added
 - sending a random gif on user request `--gif <word>`
+- the option to pass the bot token as an environment variable.
+- plugins.go to hold general plugin functions.
+- cool-down to meme, pong, and rule34 commands
+
+### Changed
+- main.go to create a struct to hold the data needed to check user request frequency.
+- logging to show the username of the requester
 
 ## Fixed
 - `invalid argument to Intn` when rule34 returns an empty list.